ex111~ex112 added

+++ ex108/teste.py
@@ -1,5 +1,5 @@
 import moeda
-price = int(input('Digite o preço: R$'))
+price = float(input('Digite o preço: R$'))
 print(f'A metade de {moeda.moeda(price)} é {moeda.moeda(moeda.metade(price))}')
 print(f'O dobro de {moeda.moeda(price)} é {moeda.moeda(moeda.dobro(price))}')
 print(f'Aumentando 10%, temos {moeda.moeda(moeda.aumentar(price, 10))}')
