+++ ex109/teste.py
@@ -1,5 +1,5 @@
 import moeda
-price = int(input('Digite o preço: R$'))
+price = float(input('Digite o preço: R$'))
 print(f'A metade de {moeda.moeda(price)} é {moeda.metade(price, True)}')
 print(f'O dobro de {moeda.moeda(price)} é {moeda.dobro(price, True)}')
 print(f'Aumentando 10%, temos {moeda.aumentar(price, 10, True)}')
