+++ ex110/teste.py
@@ -1,4 +1,4 @@
 import moeda
 
-price = int(input('Digite o preço: R$'))
+price = float(input('Digite o preço: R$'))
 moeda.resumo(price, 80, 35)
